Reduced duplication in location and class viewsets
