+++ infraohjelmointi_api/views/ProjectClassViewSet.py
@@ -1,5 +1,5 @@
 from datetime import date
-from .BaseViewSet import BaseViewSet
+from .BaseClassLocationViewSet import BaseClassLocationViewSet
 from infraohjelmointi_api.serializers.ProjectClassSerializer import (
     ProjectClassSerializer,
 )
@@ -10,43 +10,13 @@ from rest_framework.decorators import action
 from rest_framework import status
 
 
-class ProjectClassViewSet(BaseViewSet):
+class ProjectClassViewSet(BaseClassLocationViewSet):
     """
     API endpoint that allows Project Classes to be viewed or edited.
     """
 
     permission_classes = []
     serializer_class = ProjectClassSerializer
-
-    @override
-    def list(self, request, *args, **kwargs):
-        """
-        Overriden list action to get a list of ProjectClass
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Class with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-classes/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectClass instances with financial sums for projects under each class
-        """
-        year = request.query_params.get("year", date.today().year)
-        qs = self.get_queryset()
-        serializer = self.get_serializer(qs, many=True, context={"finance_year": year})
-
-        return Response(serializer.data)
 
     @override
     def get_queryset(self):
@@ -93,33 +63,6 @@ class ProjectClassViewSet(BaseViewSet):
             },
         )
         return Response(serializer.data)
-
-    def is_patch_data_valid(self, data):
-        """
-        Utility function to validate patch data sent to the custom PATCH endpoint for coordinator class finances
-        """
-        finances = data.get("finances", None)
-        if finances == None:
-            return False
-
-        parameters = finances.keys()
-        if "year" not in parameters:
-            return False
-
-        for param in parameters:
-            if param == "year":
-                continue
-            values = finances[param]
-            values_length = len(values.keys())
-
-            if not isinstance(values, dict):
-                return False
-            if values_length == 0 or values_length > 2:
-                return False
-            if "frameBudget" not in values and "budgetChange" not in values:
-                return False
-
-        return True
 
     @action(
         methods=["patch"],
