+++ infraohjelmointi_api/views/ProjectLocationViewSet.py
@@ -1,5 +1,4 @@
 from datetime import date
-from .BaseViewSet import BaseViewSet
 from infraohjelmointi_api.serializers.ProjectLocationSerializer import (
     ProjectLocationSerializer,
 )
@@ -11,45 +10,16 @@ from overrides import override
 from rest_framework.response import Response
 from rest_framework.decorators import action
 from rest_framework import status
+from .BaseClassLocationViewSet import BaseClassLocationViewSet
 
 
-class ProjectLocationViewSet(BaseViewSet):
+class ProjectLocationViewSet(BaseClassLocationViewSet):
     """
     API endpoint that allows Project Locations to be viewed or edited.
     """
 
     permission_classes = []
     serializer_class = ProjectLocationSerializer
-
-    @override
-    def list(self, request, *args, **kwargs):
-        """
-        Overriden list action to get a list of ProjectLocation
-
-            URL Query Parameters
-            ----------
-
-            year (optional) : Int
-
-            Year number to fetch Project Locations with finances starting from this year.
-            Defaults to current year.
-
-            Usage
-            ----------
-
-            project-locations/?year=<year>
-
-            Returns
-            -------
-
-            JSON
-                List of ProjectLocation instances with financial sums for projects under each location
-        """
-        year = request.query_params.get("year", date.today().year)
-        qs = self.get_queryset()
-        serializer = self.get_serializer(qs, many=True, context={"finance_year": year})
-
-        return Response(serializer.data)
 
     @override
     def get_queryset(self):
@@ -90,33 +60,6 @@ class ProjectLocationViewSet(BaseViewSet):
             },
         )
         return Response(serializer.data)
-
-    def is_patch_data_valid(self, data):
-        """
-        Utility function to validate patch data sent to the custom PATCH endpoint for coordinator location finances
-        """
-        finances = data.get("finances", None)
-        if finances == None:
-            return False
-
-        parameters = finances.keys()
-        if "year" not in parameters:
-            return False
-
-        for param in parameters:
-            if param == "year":
-                continue
-            values = finances[param]
-            values_length = len(values.keys())
-
-            if not isinstance(values, dict):
-                return False
-            if values_length == 0 or values_length > 2:
-                return False
-            if "frameBudget" not in values and "budgetChange" not in values:
-                return False
-
-        return True
 
     @action(
         methods=["patch"],
